Initial commit.
Dockerfile built on alpine. Sshd appears to work ok. Rsync is present
but so far untested. Still need to do actual launch with filesystem
mapping to test the specific remote backup usecase I have in mind.

While probably not useful in general 'as is' it should serve as an
example of how to build a modestly sized container.